Add API Client
