+++ front/client/api/Community.js
@@ -1,16 +1,14 @@
-import ApiRequest from './ApiRequest'
-import request from 'superagent'
 import RequestUrls from '../constants/RequestUrls'
+import ApiClient from '../utils/ApiClient'
 
 export const createCommunity = (name, description) => {
   const communityParams = {
     name: name,
     description: description
   }
-  return ApiRequest.post(RequestUrls.communities, communityParams)
+  return ApiClient.post(RequestUrls.communities, communityParams)
 }
 
 export const getCommunities = () => {
-  return request.get(`${Config.ApiEndPoint}${RequestUrls.communities}`)
-    .then(res => res.body)
+  return ApiClient.get(RequestUrls.communities)
 }
